+++ src/test/java/com/doozzonsa/DooZzonSaApplicationTests.java
@@ -1,0 +1,15 @@
+package com.doozzonsa;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DooZzonSaApplicationTests {
+
+    @Test
+    void contextLoads() {
+        assertThat(true).isTrue();
+    }
+}
